back/HomeLift/requirements.txt
asgiref==3.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
phonenumbers==9.0.8
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
